refactor: move main program flow to separate module
